% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{.ay_offset}
\alias{.ay_offset}
\title{Extract the academic year offset (the year AY starts) from cal_year + period}
\usage{
.ay_offset(period_index, cal_year, cal)
}
\description{
Extract the academic year offset (the year AY starts) from cal_year + period
}
\keyword{internal}
