% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.build_time_chunk}
\alias{.build_time_chunk}
\title{Build a time_chunk from a period index and calendar year}
\usage{
.build_time_chunk(period_idx, cal_year, calendar)
}
\description{
Build a time_chunk from a period index and calendar year
}
\keyword{internal}
