% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{.distance_time_chunks}
\alias{.distance_time_chunks}
\title{Compute the signed distance (in periods) between two time_chunks}
\usage{
.distance_time_chunks(x, y, cal = default_chunk_calendar())
}
\description{
Compute the signed distance (in periods) between two time_chunks
}
\keyword{internal}
