% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{.global_index}
\alias{.global_index}
\title{Convert a time_chunk to its global period index}
\usage{
.global_index(x, cal)
}
\description{
Convert a time_chunk to its global period index
}
\keyword{internal}
