% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{.global_index_to_time_chunk}
\alias{.global_index_to_time_chunk}
\title{Convert a global period index back to a time_chunk}
\usage{
.global_index_to_time_chunk(global_idx, cal)
}
\description{
Convert a global period index back to a time_chunk
}
\keyword{internal}
