% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.parse_composite_key}
\alias{.parse_composite_key}
\title{Parse a composite key string "2026-27_1_08_Fall"}
\usage{
.parse_composite_key(x, calendar)
}
\description{
Parse a composite key string "2026-27_1_08_Fall"
}
\keyword{internal}
