% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{.shift_time_chunk}
\alias{.shift_time_chunk}
\title{Shift a time_chunk vector by n periods}
\usage{
.shift_time_chunk(x, n, cal = default_chunk_calendar())
}
\description{
Shift a time_chunk vector by n periods
}
\keyword{internal}
