% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.two_digit_year}
\alias{.two_digit_year}
\title{Convert a 2-digit year to a 4-digit year}
\usage{
.two_digit_year(yy)
}
\description{
Follows standard R convention: 00-49 -> 2000-2049, 50-99 -> 1950-1999.
}
\keyword{internal}
