% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is_time_chunk}
\alias{is_time_chunk}
\title{Test if an object is a time_chunk}
\usage{
is_time_chunk(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{time_chunk}, \code{FALSE} otherwise.
}
\description{
Test if an object is a time_chunk
}
\examples{
use_chunk_preset("us_semester")
is_time_chunk(time_chunk("fa26"))  # TRUE
is_time_chunk("fa26")              # FALSE
}
