% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{print.time_chunk}
\alias{print.time_chunk}
\title{Print method for time_chunk vectors}
\usage{
\method{print}{time_chunk}(x, ...)
}
\arguments{
\item{x}{A \code{time_chunk} vector.}

\item{...}{Ignored.}
}
\description{
Print method for time_chunk vectors
}
