% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{validate_time_chunk}
\alias{validate_time_chunk}
\title{Validate a time_chunk object}
\usage{
validate_time_chunk(x)
}
\arguments{
\item{x}{A \code{time_chunk} object.}
}
\value{
\code{x} invisibly, or errors with a descriptive message.
}
\description{
Validate a time_chunk object
}
\keyword{internal}
